requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=1.10.7
pytest>=7.0.0
python-dotenv>=1.0.0
langgraph>=0.0.37
tqdm>=4.65.0
